requirements.txt
langchain-openai>=0.2.0
guardrails-ai>=0.4.0
python-dotenv>=1.0
